pull destinations ref to formsConfigs + add lts layers back #45
