+++ analysis/js/sidebar/formsConfigs.js
@@ -15,5 +15,12 @@ const ltsFilters = {
     ]
 }
 
+// order matters - mapbox prioritizes layers by order & we want analysis on top
+const specialDestinationLayers = {
+    school: ['schools-combined', 'school', 'school-ipd'],
+    trails: ['trail-access', 'trails', 'trails-ipd'],
+    transit: ['trolley', 'passenger-rail', 'bus', 'transit', 'transit-ipd'],
+    priority: ['priority', 'priority-ipd']
+}
 
-export { ltsFilters }
+export { ltsFilters, specialDestinationLayers }
